requirements.txt
#Install manually
#torch

transformers
matplotlib
pillow
huggingface-hub
gradio==5.9.0
numpy==1.26.4
onnxruntime
rembg